Copy and delete files in virtual directories
fs::rename doesn't work correctly on virtual Linux directories.
