+++ crates/cargo-lambda-metadata/Cargo.toml
@@ -20,3 +20,7 @@ cargo_metadata = "0.14.2"
 miette = "4.3.0"
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.79"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+libc = "0.2.123"
+remove_dir_all = "0.7.0"
